# core/vm: JIT segmentation