--- a/core/vm/stack.go
+++ b/core/vm/stack.go
@@ -42,6 +42,9 @@ func (st *stack) push(d *big.Int) {
 	//st.data = append(st.data, stackItem)
 	st.data = append(st.data, d)
 }
+func (st *stack) pushN(ds ...*big.Int) {
+	st.data = append(st.data, ds...)
+}
 
 func (st *stack) pop() (ret *big.Int) {
 	ret = st.data[len(st.data)-1]