This PR adds JIT optimisations through segmentation and pre calculating segmentation gas chunks. 

Currently implements:
* multi-pushes (`PUSH 0x1 PUSH 0x1 ...`)
* static jumps (` PUSH 0x1 JUMP`)

This PR also changes how the JIT executes instructions.  Previously the JIT was responsible for figuring out how an instruction needed to be executed. This has changed to a generic model where the JIT calls the `type programInstruction interface { ... }` which puts the program instruction in charge of the implementation specific details such as incrementing the `pc`, executing the instruction functions, stack validations, etc. This generic model helps when adding segments because it requires control over the order in which the JIT executes instructions.