@obscuren 2015-10-06
```
 tests git:(jit-vm-optimisations) $ benchcmp ~/before.txt ~/after.txt
benchmark                         old ns/op     new ns/op     delta
BenchmarkVmAckermann32Tests-4     6508880       5535155       -14.96%
BenchmarkVmFibonacci16Tests-4     48406689      36364750      -24.88%
```