@obscuren 2015-10-07
Some early review requested