@Gustav-Simonsson 2015-10-09
LGTM, maybe add some more inline comments about how segments work and why they are cheaper to execute (how they are compiled) compared to regular VM execution of instructions.